requirements.txt
flask==3.0.0
flask-cors==4.0.0
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
mysql-connector-python==8.2.0
python-dotenv==1.0.0
pandas==2.1.4
sqlalchemy==2.0.23
openpyxl==3.1.2
xlsxwriter==3.1.9

